+++ Example/PPEventRegistryTest tvOS/ViewController.swift
@@ -11,19 +11,25 @@ import PPEventRegistryAPI
 
 class ViewController: UIViewController {
 
-    let eventRegistryAPI = PPEventRegistryAPI()
+    let api = PPEventRegistryAPI()
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        eventRegistryAPI.login("", password: "") { (error) in
+        api.login("", password: "") { error in
             print("login with error: \(error)")
-            self.eventRegistryAPI.getEvent(withID: 4488701, completionHandler: { (event, error) in
-                print("get event error \(event)")
-            })
-            self.eventRegistryAPI.getRecentArticles({ (aricles, error) in
-                print("articles \(aricles), error \(error)")
-            })
+            self.api.getEvent(withID: 44808701) { event, error in
+                print("get event \(event), error \(error)")
+            }
+            self.api.getRecentArticles(marker: nil) { marker, aricles, error in
+                print("articles, request1 \(marker): \(aricles), error \(error)")
+
+                if let marker = marker {
+                    self.api.getRecentArticles(marker: marker) { marker, aricles, error in
+                        print("articles, request2 \(marker): \(aricles), error \(error)")
+                    }
+                }
+            }
         }
     }