+++ Example/PPEventRegistryTest watchOS Extension/InterfaceController.swift
@@ -12,7 +12,7 @@ import PPEventRegistryAPI
 
 class InterfaceController: WKInterfaceController {
 
-    let eventRegistryAPI = PPEventRegistryAPI()
+    let api = PPEventRegistryAPI()
 
     override func awake(withContext context: Any?) {
         super.awake(withContext: context)
@@ -24,14 +24,20 @@ class InterfaceController: WKInterfaceController {
         // This method is called when watch view controller is about to be visible to user
         super.willActivate()
         
-        eventRegistryAPI.login("", password: "") { (error) in
+        api.login("", password: "") { error in
             print("login with error: \(error)")
-            self.eventRegistryAPI.getEvent(withID: 4488701, completionHandler: { (event, error) in
-                print("get event error \(event)")
-            })
-            self.eventRegistryAPI.getRecentArticles({ (aricles, error) in
-                print("articles \(aricles), error \(error)")
-            })
+            self.api.getEvent(withID: 44808701) { event, error in
+                print("get event \(event), error \(error)")
+            }
+            self.api.getRecentArticles(marker: nil) { marker, aricles, error in
+                print("articles, request1 \(marker): \(aricles), error \(error)")
+
+                if let marker = marker {
+                    self.api.getRecentArticles(marker: marker) { marker, aricles, error in
+                        print("articles, request2 \(marker): \(aricles), error \(error)")
+                    }
+                }
+            }
         }
     }